+++ src/realm/tokenizer.cpp
@@ -41,6 +41,32 @@ std::set<std::string> Tokenizer::get_all_tokens()
     return tokens;
 }
 
+TokenInfoMap Tokenizer::get_token_info()
+{
+    TokenInfoMap info;
+    size_t num_tokens = 0;
+    while (next()) {
+        auto it = info.find(std::string(get_token()));
+        if (it == info.end()) {
+            info.emplace(std::string(get_token()), TokenInfo(num_tokens, get_range()));
+        }
+        else {
+            TokenInfo& i = it->second;
+            i.positions.emplace_back(num_tokens);
+            i.ranges.emplace_back(get_range());
+            i.weight *= 2;
+            i.frequency += (1 / i.weight);
+        }
+        num_tokens++;
+    }
+    for (auto& it : info) {
+        TokenInfo& i = it.second;
+        double coeff = (0.5 * i.positions.size() / num_tokens) + 0.5;
+        i.weight = i.frequency * coeff;
+    }
+    return info;
+}
+
 class DefaultTokenizer : public Tokenizer {
 public:
     bool next() override;
@@ -112,10 +138,14 @@ bool DefaultTokenizer::next()
         }
 
         if (is_alnum) {
+            if (state == searching) {
+                m_start = m_cur_pos - m_start_pos;
+            }
             state = building;
         }
         else {
             if (state == building) {
+                m_end = m_cur_pos - m_start_pos;
                 state = finished;
             }
         }
@@ -131,3 +161,58 @@ std::unique_ptr<Tokenizer> Tokenizer::get_instance()
 }
 
 } // namespace realm
+#ifdef TOKENIZER_UNITTEST
+
+// compile: g++ -DTOKENIZER_UNITTEST=1 -I.. --std=c++17 -g -o test_tokenizer tokenizer.cpp
+
+#include <cassert>
+#include <iostream>
+#include <fstream>
+#include <chrono>
+
+using namespace std::chrono;
+
+char buffer[256 * 256];
+
+int main(int argc, const char* argv[])
+{
+    auto tok = realm::Tokenizer::get_instance();
+    tok->reset("to be or not to be");
+    auto tokens = tok->get_all_tokens();
+    assert(tokens.size() == 4);
+    tok->reset("To be or not to be");
+    realm::TokenInfoMap info = tok->get_token_info();
+    assert(info.size() == 4);
+    realm::TokenInfo& i(info["to"]);
+    assert(i.positions.size() == 2);
+    assert(i.positions[0] == 0);
+    assert(i.positions[1] == 4);
+    assert(i.ranges.size() == 2);
+    assert(i.ranges[0].first == 0);
+    assert(i.ranges[0].second == 2);
+    assert(i.ranges[1].first == 13);
+    assert(i.ranges[1].second == 15);
+    tok->reset("Jeg gik mig over sø og land");
+    info = tok->get_token_info();
+    assert(info.size() == 7);
+    realm::TokenInfo& j(info["sø"]);
+    assert(j.ranges[0].first == 17);
+    assert(j.ranges[0].second == 20);
+
+    if (argc > 1) {
+        std::ifstream istr(argv[1]);
+        istr.read(buffer, sizeof(buffer));
+        std::string_view text(buffer, istr.gcount());
+
+        tok->reset(text);
+        auto t1 = steady_clock::now();
+        auto tokens = tok->get_all_tokens();
+        auto t2 = steady_clock::now();
+        tok->reset(text);
+        auto info = tok->get_token_info();
+        auto t3 = steady_clock::now();
+        std::cout << "tokenize: " << duration_cast<microseconds>(t2 - t1).count() << " us" << std::endl;
+        std::cout << "info: " << duration_cast<microseconds>(t3 - t2).count() << " us" << std::endl;
+    }
+}
+#endif